Start the StorPool OpenStack integration suite.
Start gathering up the StorPool Cinder and Nova driver files for
the various distributions, flavors and versions of OpenStack that
we are likely to encounter.  For the present, go with the latest
upstream versions for OpenStack Juno and Kilo.

Also start documenting the process of setting up the StorPool
drivers on an existing OpenStack cluster (and an existing StorPool
cluster, too).

ObQuote:	"A beginning is a very delicate time."