Remove the non-OpenStack drivers.
Now that the driver update mechanism is much more streamlined,
there is no real sense in maintaining our copies of all the drivers
for all distributions, releases, branches, etc.  It is much easier and
less error-prone to regenerate the drivers every time.

ObQuote:	"It's my way, my way or the highway"