Rebuild the Kolla Docker containers with our changes
Introduce the `kolla_rebuild` module and the kolla/docker/Dockerfile.j2
Jinja template. Rebuild the cinder-volume and nova-compute upstream Kolla
containers so that the StorPool Cinder driver is fully patched and
the /var/spool/openstack-storpool directory, permissions, and user groups
are configured correctly for the respective OpenStack service account.

Add the "OpenStack Kolla integration" section to the documentation.