Switch to SPDX license identifiers
Add the Apache 2.0 license text in the LICENSES/ subdirectory.

Drop the license blurbs from the files that had them, replace them with
SPDX-License-Identifier lines.

Add SPDX-FileCopyrightText and SPDX-License-Identifier tags to pretty much
all of the source files.

Add a .reuse/dep5 file that lists the ones that we do not want or cannot
modify to specify copyright and license information.